rename: 공통 회원가입 뷰 common 파일로 이동
