+++ lib/features/auth/signup/views/common/signup_screen_choose.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:cogo/features/auth/signup/views_model/signup_choose_view_model.dart';
+import 'package:cogo/features/auth/signup/views_model/common/signup_choose_view_model.dart';
 import 'package:cogo/common/widgets/custom_appbar.dart';
 
 class SignupScreenChoose extends StatelessWidget {
