+++ lib/features/auth/signup/views_model/common/agreement_view_model.dart
