+++ lib/route/routes.dart
@@ -4,10 +4,10 @@ import 'package:cogo/features/home/views/home_screen.dart';
 import 'package:cogo/features/cogo/views/cogo_screen.dart';
 import 'package:cogo/features/mypage/views/mypage_screen.dart';
 import 'package:cogo/common/navigator/view/bottom_navigation_bar.dart';
-import 'package:cogo/features/auth/signup/views/agreement_screen.dart';
-import 'package:cogo/features/auth/signup/views/signup_screen_phone.dart';
-import 'package:cogo/features/auth/signup/views/signup_screen_name.dart';
-import 'package:cogo/features/auth/signup/views/signup_screen_choose.dart';
+import 'package:cogo/features/auth/signup/views/common/agreement_screen.dart';
+import 'package:cogo/features/auth/signup/views/common/signup_screen_phone.dart';
+import 'package:cogo/features/auth/signup/views/common/signup_screen_name.dart';
+import 'package:cogo/features/auth/signup/views/common/signup_screen_choose.dart';
 import 'package:cogo/features/auth/signup/views/mentor/interest_selection_screen.dart';
 import 'package:cogo/features/auth/signup/views/mentor/club_selection_screen.dart';
 import 'package:cogo/features/auth/signup/views/mentor/mentor_info_screen.dart';
